Added collision system, kirby
Also some testing is included

+++ Definitions.h
@@ -1,3 +1,3 @@
 #pragma once
 
-typedef int numero;
+typedef double numero;
